# Conflicts:
#	compose/docker-compose-prod.yml
#	ontrack-service/src/main/java/net/nemerosa/ontrack/service/security/RolesServiceImpl.kt
#	ontrack-service/src/test/java/net/nemerosa/ontrack/service/security/SecurityServiceIT.java

+++ ontrack-extension-indicators/src/main/java/net/nemerosa/ontrack/extension/indicators/model/IndicatorTypeServiceImpl.kt
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.node.NullNode
 import net.nemerosa.ontrack.common.getOrNull
 import net.nemerosa.ontrack.extension.indicators.acl.IndicatorTypeManagement
 import net.nemerosa.ontrack.model.Ack
+import net.nemerosa.ontrack.model.security.GlobalSettings
 import net.nemerosa.ontrack.model.security.SecurityService
 import net.nemerosa.ontrack.model.support.StorageService
 import org.springframework.stereotype.Service
@@ -13,10 +14,10 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional
 class IndicatorTypeServiceImpl(
-        private val indicatorCategoryService: IndicatorCategoryService,
-        private val indicatorValueTypeService: IndicatorValueTypeService,
-        private val storageService: StorageService,
-        private val securityService: SecurityService
+    private val indicatorCategoryService: IndicatorCategoryService,
+    private val indicatorValueTypeService: IndicatorValueTypeService,
+    private val storageService: StorageService,
+    private val securityService: SecurityService
 ) : IndicatorTypeService, IndicatorCategoryListener {
 
     init {
@@ -41,29 +42,29 @@ class IndicatorTypeServiceImpl(
         }.mapNotNull {
             fromStorage<Any, Any>(it)
         }.sortedWith(
-                compareBy(
-                        { it.category.name },
-                        { it.name }
-                )
+            compareBy(
+                { it.category.name },
+                { it.name }
+            )
         )
     }
 
     override fun findTypeById(typeId: String): IndicatorType<*, *>? =
-            storageService.retrieve(STORE, typeId, StoredIndicatorType::class.java)
-                    .getOrNull()
-                    ?.let { fromStorage<Any, Any>(it) }
+        storageService.retrieve(STORE, typeId, StoredIndicatorType::class.java)
+            .getOrNull()
+            ?.let { fromStorage<Any, Any>(it) }
 
     override fun getTypeById(typeId: String): IndicatorType<*, *> =
-            findTypeById(typeId) ?: throw IndicatorTypeNotFoundException(typeId)
+        findTypeById(typeId) ?: throw IndicatorTypeNotFoundException(typeId)
 
     override fun findByCategory(category: IndicatorCategory): List<IndicatorType<*, *>> {
         return findAll().filter {
             it.category.id == category.id
         }.sortedWith(
-                compareBy(
-                        { it.category.name },
-                        { it.name }
-                )
+            compareBy(
+                { it.category.name },
+                { it.name }
+            )
         )
     }
 
@@ -73,14 +74,15 @@ class IndicatorTypeServiceImpl(
         return if (category != null && valueType != null) {
             val valueConfig = valueType.fromConfigStoredJson(stored.valueConfig)
             IndicatorType(
-                    id = stored.id,
-                    category = category,
-                    name = stored.name,
-                    link = stored.link,
-                    valueType = valueType,
-                    valueConfig = valueConfig,
-                    source = stored.source,
-                    computed = stored.computed
+                id = stored.id,
+                category = category,
+                name = stored.name,
+                link = stored.link,
+                valueType = valueType,
+                valueConfig = valueConfig,
+                source = stored.source,
+                computed = stored.computed,
+                deprecated = stored.deprecated
             )
         } else {
             null
@@ -98,14 +100,15 @@ class IndicatorTypeServiceImpl(
     }
 
     override fun <T, C> createType(
-            id: String,
-            category: IndicatorCategory,
-            name: String,
-            link: String?,
-            valueType: IndicatorValueType<T, C>,
-            valueConfig: C,
-            source: IndicatorSource?,
-            computed: Boolean
+        id: String,
+        category: IndicatorCategory,
+        name: String,
+        link: String?,
+        valueType: IndicatorValueType<T, C>,
+        valueConfig: C,
+        source: IndicatorSource?,
+        computed: Boolean,
+        deprecated: String?
     ): IndicatorType<T, C> {
         securityService.checkGlobalFunction(IndicatorTypeManagement::class.java)
         val type = findTypeById(id)
@@ -113,27 +116,46 @@ class IndicatorTypeServiceImpl(
             throw IndicatorTypeIdAlreadyExistsException(id)
         } else {
             return updateType(
-                    id,
-                    category,
-                    name,
-                    link,
-                    valueType,
-                    valueConfig,
-                    source,
-                    computed
+                id = id,
+                category = category,
+                name = name,
+                link = link,
+                valueType = valueType,
+                valueConfig = valueConfig,
+                source = source,
+                computed = computed,
+                deprecated = deprecated
             )
         }
     }
 
-    override fun deleteType(id: String): Ack {
+    override fun deleteType(id: String, force:Boolean): Ack {
         securityService.checkGlobalFunction(IndicatorTypeManagement::class.java)
+        if (force) securityService.checkGlobalFunction(GlobalSettings::class.java)
         val type = findTypeById(id)
         return if (type != null) {
-            listeners.forEach { it.onTypeDeleted(type) }
-            storageService.delete(STORE, id)
-            Ack.OK
+            if (force || type.source == null || !type.deprecated.isNullOrBlank()) {
+                listeners.forEach { it.onTypeDeleted(type) }
+                storageService.delete(STORE, id)
+                Ack.OK
+            } else {
+                Ack.NOK
+            }
         } else {
             Ack.NOK
+        }
+    }
+
+    override fun deprecateType(id: String, deprecated: String?) {
+        securityService.checkGlobalFunction(IndicatorTypeManagement::class.java)
+        val existing = storageService.retrieve(STORE, id, StoredIndicatorType::class.java).getOrNull()
+        if (existing != null) {
+            val deprecatedEntry = existing.withDeprecated(deprecated)
+            storageService.store(
+                STORE,
+                id,
+                deprecatedEntry
+            )
         }
     }
 
@@ -142,66 +164,82 @@ class IndicatorTypeServiceImpl(
         val category = indicatorCategoryService.getCategory(input.category)
         val valueType = indicatorValueTypeService.getValueType<Any, Any>(input.valueType.id)
         val valueConfig = valueType.toConfigStoredJson(
-                valueType.fromConfigForm(input.valueType.data ?: NullNode.instance)
+            valueType.fromConfigForm(input.valueType.data ?: NullNode.instance)
         )
         val stored = StoredIndicatorType(
-                id = input.id,
-                category = category.id,
-                name = input.name,
-                link = input.link,
-                valueType = valueType.id,
-                valueConfig = valueConfig,
-                source = null,
-                computed = false
+            id = input.id,
+            category = category.id,
+            name = input.name,
+            link = input.link,
+            valueType = valueType.id,
+            valueConfig = valueConfig,
+            source = null,
+            computed = false,
+            deprecated = input.deprecated?.takeIf { it.isNotBlank() }
         )
         storageService.store(
-                STORE,
-                input.id,
-                stored
+            STORE,
+            input.id,
+            stored
         )
         return getTypeById(input.id)
     }
 
     override fun <T, C> updateType(
-            id: String,
-            category: IndicatorCategory,
-            name: String,
-            link: String?,
-            valueType: IndicatorValueType<T, C>,
-            valueConfig: C,
-            source: IndicatorSource?,
-            computed: Boolean
+        id: String,
+        category: IndicatorCategory,
+        name: String,
+        link: String?,
+        valueType: IndicatorValueType<T, C>,
+        valueConfig: C,
+        source: IndicatorSource?,
+        computed: Boolean,
+        deprecated: String?
     ): IndicatorType<T, C> {
         securityService.checkGlobalFunction(IndicatorTypeManagement::class.java)
         val stored = StoredIndicatorType(
-                id = id,
-                category = category.id,
-                name = name,
-                link = link,
-                valueType = valueType.id,
-                valueConfig = valueType.toConfigStoredJson(valueConfig),
-                source = source,
-                computed = computed
+            id = id,
+            category = category.id,
+            name = name,
+            link = link,
+            valueType = valueType.id,
+            valueConfig = valueType.toConfigStoredJson(valueConfig),
+            source = source,
+            computed = computed,
+            deprecated = deprecated?.takeIf { it.isNotBlank() }
         )
         storageService.store(
-                STORE,
-                id,
-                stored
+            STORE,
+            id,
+            stored
         )
         @Suppress("UNCHECKED_CAST")
         return getTypeById(id) as IndicatorType<T, C>
     }
 
     private class StoredIndicatorType(
-            val id: String,
-            val category: String,
-            val name: String,
-            val link: String?,
-            val valueType: String,
-            val valueConfig: JsonNode,
-            val source: IndicatorSource?,
-            val computed: Boolean
-    )
+        val id: String,
+        val category: String,
+        val name: String,
+        val link: String?,
+        val valueType: String,
+        val valueConfig: JsonNode,
+        val source: IndicatorSource?,
+        val computed: Boolean,
+        val deprecated: String?
+    ) {
+        fun withDeprecated(deprecated: String?) = StoredIndicatorType(
+            id,
+            category,
+            name,
+            link,
+            valueType,
+            valueConfig,
+            source,
+            computed,
+            deprecated
+        )
+    }
 
     companion object {
         private val STORE: String get() = IndicatorType::class.java.name
